OTUS-5: Added localization and horizontal orientation for film details activity
