+++ app/src/main/java/ru/evilsnow/otus/fcatalog/MainActivity.kt
@@ -1,6 +1,6 @@
 package ru.evilsnow.otus.fcatalog
 
-import android.content.Context
+import android.app.Dialog
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -8,12 +8,14 @@ import android.view.*
 import android.widget.*
 import androidx.core.view.get
 import ru.evilsnow.otus.fcatalog.dao.FilmsDao
-import ru.evilsnow.otus.fcatalog.model.FilmItem
+import ru.evilsnow.otus.fcatalog.model.FilmItemsAdapter
+import ru.evilsnow.otus.fcatalog.ui.ExitDialog
 
 class MainActivity : AppCompatActivity() {
 
     private lateinit var mFilmsDao: FilmsDao
     private var lastSelectedFilmPosition: Int? = null
+    private var mConfirmExitDialog: Dialog? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -76,6 +78,13 @@ class MainActivity : AppCompatActivity() {
         return super.onOptionsItemSelected(item)
     }
 
+    override fun onBackPressed() {
+        if (mConfirmExitDialog == null) {
+            mConfirmExitDialog = ExitDialog(this) { super.onBackPressed() }
+        }
+        mConfirmExitDialog!!.show()
+    }
+
     private fun shareWithFriends() {
         val shareIntent = Intent().apply {
             action = Intent.ACTION_SEND
@@ -83,46 +92,6 @@ class MainActivity : AppCompatActivity() {
             putExtra(Intent.EXTRA_TEXT, "Hello. New cool app for download: https://github.com/evilsnow-ru/films-catalog")
         }
         startActivity(Intent.createChooser(shareIntent, "Share with friends"))
-    }
-
-    class FilmItemsAdapter(
-        private val mContext: Context,
-        private val filmsArray: Array<FilmItem>,
-        private val mSelectedIndex: Int?,
-        private val detailsCallback: (Int, Int) -> Unit
-    ) : BaseAdapter() {
-
-        private val mInflater: LayoutInflater = LayoutInflater.from(mContext)
-
-        override fun getView(position: Int, convertView: View?, parent: ViewGroup?): View {
-            val view = mInflater.inflate(R.layout.film_list_item, parent, false)
-            val filmItem = getItem(position)
-
-            val titleView = view.findViewById<TextView>(R.id.filmTitle)
-            titleView.text = filmItem.title
-
-            mSelectedIndex?.let {
-                if (it == position) {
-                    titleView.setTextColor(mContext.getColor(R.color.selectedTitle))
-                }
-            }
-
-            view.findViewById<ImageView>(R.id.filmTitleThumbnail).setImageResource(filmItem.image)
-
-            view.findViewById<Button>(R.id.filmDetailsBtn).setOnClickListener {
-                titleView.setTextColor(mContext.getColor(R.color.selectedTitle))
-                detailsCallback.invoke(position, filmItem.id)
-            }
-
-            return view
-        }
-
-        override fun getItem(position: Int): FilmItem = filmsArray[position]
-
-        override fun getItemId(position: Int): Long = filmsArray[position].id.toLong()
-
-        override fun getCount(): Int = filmsArray.size
-
     }
 
 }
