+++ app/src/main/res/layout/dialog_exit.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <TextView
+        android:id="@+id/exit_question_text"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/exit_question"
+        android:textSize="24sp"
+        android:textColor="@color/black"
+        android:layout_margin="5dp"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/exit_negative_button"
+        style="@style/ExitConfirmDialogButton"
+        android:text="@string/dialog_exit_negative_button"
+        app:layout_constraintEnd_toStartOf="@+id/exit_positive_button"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/exit_question_text" />
+
+    <Button
+        android:id="@+id/exit_positive_button"
+        style="@style/ExitConfirmDialogButton"
+        android:text="@string/dialog_exit_positive_button"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@id/exit_negative_button"
+        app:layout_constraintTop_toBottomOf="@id/exit_question_text" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
