+++ app/src/main/res/layout/film_list_item.xml
@@ -1,41 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
-    android:layout_height="wrap_content" xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content">
 
     <ImageView
         android:id="@+id/filmTitleThumbnail"
         android:layout_width="100dp"
         android:layout_height="100dp"
-        android:contentDescription="@string/title_thumbnail_str"
         android:layout_margin="5dp"
+        android:contentDescription="@string/title_thumbnail_str"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/filmTitle"
+        style="@style/MainFilmTitle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        android:textColor="@color/black"
-        android:layout_margin="5dp"
-        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toEndOf="@id/filmTitleThumbnail"
+        app:layout_constraintTop_toTopOf="parent"
         tools:text="Film title" />
 
     <Button
         android:id="@+id/filmDetailsBtn"
         android:layout_width="0dp"
         android:layout_height="45dp"
-        android:text="@string/details_button_text"
         android:layout_margin="5dp"
-        app:layout_constraintStart_toEndOf="@id/filmTitleThumbnail"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/filmTitle"
+        android:text="@string/details_button_text"
         app:layout_constraintBottom_toBottomOf="parent"
-        />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@id/filmTitleThumbnail"
+        app:layout_constraintTop_toBottomOf="@id/filmTitle" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
