+++ app/src/main/res/menu/activity_main_menu.xml
@@ -4,7 +4,7 @@
 
     <item
         android:id="@+id/share_menu_item"
-        android:icon="@android:drawable/ic_menu_share"
+        android:icon="@drawable/ic_share_24px"
         android:title="@string/mi_activity_main_share"
         app:showAsAction="ifRoom|withText"/>
